pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "wallpaper-gen"
version = "0.1.0"
description = "AI Wallpaper Generation Pipeline with Streamlit UI"
authors = [
    {name = "benja"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "streamlit>=1.30.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "Pillow>=10.0.0",
    "safetensors>=0.4.0",
    "realesrgan>=0.3.0",
    "basicsr>=1.4.2",
    "facexlib>=0.3.0",
    "gfpgan>=1.3.8",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[tool.poetry.dependencies]
torch = {version = ">=2.1.0", source = "pytorch-cu124"}
torchvision = {version = ">=0.16.0", source = "pytorch-cu124"}
